Refactor: WebView 디바이스 감지 로직을 별도 유틸리티로 분리하여 hook 간소화
